pyproject.toml
[tool.poetry]
name = "koko-devops"
version = "0.1.0"
description = ""
authors = ["EduhG <edwinnyangena@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.1.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.6.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
